added login, signup, and main tabs routes. Setup backend api calls

+++ server/controllers/usersController.js
@@ -1,6 +1,6 @@
 const asyncHandler = require("express-async-handler");
 const user_model = require("../models/users");
-const user_middleware = require("../middleware/usersMiddleware");
+const { generate_auth_tokens } = require("./auth");
 const bcrypt = require("bcrypt");
 
 // displays user registrations page
@@ -10,87 +10,110 @@ exports.user_register_page = asyncHandler(async (req, res, next) => {
     .json({ message: "Not implemented: user registration GET page" });
 });
 
-// creates a new user in the DB
-exports.user_register = asyncHandler(async (req, res, next) => {
-  let { username, password } = req.body;
-  let existingUser = await user_model.get_user_by_username(username);
-
-  if (!existingUser) {
-    let hashedPassword = await user_middleware.hash_password(password);
-    await user_model.create_user(username, hashedPassword).catch((error) => {
-      console.log(error);
-      res.status(500).send({ error: error });
+exports.verifyCredentials = async (username, phone) => {
+  try {
+    let existingUsername = await user_model.get_user_by_username(username);
+    let existingPhoneNumber = await user_model.check_existing_phone(phone);
+    return {
+      usernameExists: Boolean(existingUsername),
+      phoneNumberExists: Boolean(existingPhoneNumber),
+    };
+  } catch (err) {
+    console.log(err);
+    return err;
+  }
+};
+exports.user_verify_unique = asyncHandler(async (req, res, next) => {
+  let { username, phoneNumber, password } = req.body.loginInfo;
+  console.log(req.body.loginInfo);
+  let { usernameExists, phoneNumberExists } = await this.verifyCredentials(
+    username,
+    phoneNumber
+  );
+  if (!usernameExists && !phoneNumberExists) {
+    let passwordHash = await bcrypt.hash(password, 8);
+    res.status(200).json({
+      message: "Verified username and phone number!",
+      loginInfo: { username, phoneNumber, password: passwordHash },
     });
-
-    res.status(200).json({ message: "Registered successfully" });
   } else {
-    res.status(401).send({ error: "This username is taken" });
+    res.status(401).json({
+      usernameExists,
+      phoneNumberExists,
+    });
   }
 });
 
 // displays update user information page
 exports.user_update_page = asyncHandler(async (req, res, next) => {
-  res.send("Not implemented: user update GET page");
+  res.json("Not implemented: user update GET page");
 });
 
 // update user information
 exports.user_update_username = asyncHandler(async (req, res, next) => {
-  let { username, password, newUsername } = req.body;
-  let authenticated = await user_middleware.verify_user(username, password);
-  if (authenticated.error) {
-    if (
-      authenticated.error == "Invalid username" ||
-      authenticated.error == "Incorrect password"
-    ) {
-      res.status(401).send({ error: "Could not authenticate user" });
+  if (req.decoded && req.decoded.user_id == req.params.id) {
+    let { newUsername } = req.body;
+    if (!newUsername || req.decoded.username == newUsername) {
+      res.status(401).json({ error: "Invalid new username" });
     } else {
-      res.status(500).send(authenticated);
+      let existingUser = await user_model.get_user_by_username(newUsername);
+
+      if (!existingUser) {
+        let update = await user_model
+          .update_username(req.decoded.user_id, newUsername)
+          .catch((err) => {
+            console.log(err);
+            res.status(500).json({ error: err });
+          });
+
+        let accessTokens = generate_auth_tokens(
+          req.decoded.user_id,
+          newUsername
+        );
+
+        res
+          .status(200)
+          .json({ ...accessTokens, message: "Username updated successfully" });
+      } else {
+        res.status(401).json({ error: "Username taken" });
+      }
     }
   } else {
-    let existingUser = await user_model.get_user_by_username(newUsername);
-    if (!existingUser) {
-      let update = await user_model
-        .update_username(req.params.id, newUsername)
-        .catch((err) => {
-          res.status(500).send({ error: err });
-        });
-      res.status(200).send();
-    }
+    res.status(401).json({ error: "Not authorized" });
   }
 });
 
 exports.user_update_password = asyncHandler(async (req, res, next) => {
-  let { username, password, newPassword } = req.body;
-  let authenticated = await user_middleware.verify_user(username, password);
-
-  if (authenticated.error) {
-    if (
-      authenticated.error == "Invalid username" ||
-      authenticated.error == "Incorrect password"
-    ) {
-      res.status(401).send({ error: "Could not authenticate user" });
+  if (req.decoded && req.decoded.user_id == req.params.id) {
+    let { newPassword } = req.body;
+    if (!newPassword) {
+      res.status(401).json({ error: "Invalid new password" });
     } else {
-      res.status(500).send(authenticated);
+      let passwordHash = await bcrypt.hash(newPassword, 8);
+      await user_model
+        .update_password(req.params.id, passwordHash)
+        .catch((err) => {
+          res.status(500).json({ error: err });
+        });
+      res.status(200).json();
     }
   } else {
-    let passwordHash = await user_middleware.hash_password(newPassword);
-
-    let update = await user_model
-      .update_password(req.params.id, passwordHash)
-      .catch((err) => {
-        res.status(500).send({ error: err });
-      });
-    res.status(200).send();
+    res.status(401).json({ error: "Not authorized" });
   }
 });
+
 // get user by id (for account info page)
 exports.user_get_by_id = asyncHandler(async (req, res, next) => {
-  let user = await user_model.get_user_by_id(req.params.id);
+  if (req.decoded && req.decoded.user_id) {
+    let user = await user_model.get_user_by_id(req.decoded.user_id);
 
-  if (user) {
-    res.status(200).json({ user: user });
+    if (user) {
+      res.status(200).json({ user: user });
+    } else {
+      res.status(404).json({ error: "User not found" });
+    }
   } else {
-    res.status(404).json({ error: "user not found" });
+    res.status(401).json({ error: "Not authorized" });
   }
 });
 
@@ -100,40 +123,19 @@ exports.user_get_by_username = asyncHandler(async (req, res, next) => {
   if (user) {
     res.status(200).json({ user: user });
   } else {
-    res.status(404).json({ error: "user not found" });
+    res.status(404).json({ error: "User not found" });
   }
 });
 
 // display login page
 exports.user_login_page = asyncHandler(async (req, res, next) => {
-  res.send("Not implemented: user login GET page");
-});
-
-// verify user login info
-exports.user_login = asyncHandler(async (req, res, next) => {
-  let { username, password } = req.body;
-
-  let authenticated = await user_middleware.verify_user(username, password);
-
-  if (authenticated.error) {
-    if (
-      authenticated.error == "Incorrect password" ||
-      authenticated.error == "Invalid username"
-    ) {
-      res.status(401).send(authenticated);
-    } else {
-      res.status(500).send(authenticated);
-    }
-  } else {
-    res.status(200).json({ message: "Login successful" });
-  }
+  res.json("Not implemented: user login GET page");
 });
 
 // list users
 exports.users_list = asyncHandler(async (req, res, next) => {
   let allUsers = user_model.list_users();
   allUsers.then((users) => {
-    console.log(users);
     res.status(200).json({ users: users });
   });
 });
@@ -141,16 +143,20 @@ exports.users_list = asyncHandler(async (req, res, next) => {
 // query users by username autocomplete
 exports.users_query_username = asyncHandler(async (req, res, next) => {
   let queryTerm = req.body.queryTerm;
-  let usernames = await user_model.search_usernames(queryTerm).catch((err) => {
+  let users = await user_model.search_users(queryTerm).catch((err) => {
     res.status(401).json({ message: "Something went wrong" });
   });
-  res.status(200).json({ users: usernames });
+  res.status(200).json({ users: users });
 });
 
 // delete user
 exports.user_delete = asyncHandler(async (req, res, next) => {
-  let deleted = await user_model.delete_user(req.params.id).catch((err) => {
-    res.status(500).send({ error: err });
-  });
-  res.status(200).json({ message: "Successfully deleted" });
+  if (req.decoded && req.decoded.user_id == req.params.id) {
+    let deleted = await user_model.delete_user(req.params.id).catch((err) => {
+      res.status(500).json({ error: err });
+    });
+    res.status(200).json({ message: "Successfully deleted" });
+  } else {
+    res.status(401).json({ error: "Not authorized" });
+  }
 });
